+++ sqlx-core/src/mysql/types/chrono.rs
@@ -1,5 +1,8 @@
 use bytes::Buf;
-use chrono::{DateTime, Datelike, Local, NaiveDate, NaiveDateTime, NaiveTime, Timelike, Utc};
+use chrono::{
+    DateTime, Datelike, FixedOffset, Local, NaiveDate, NaiveDateTime, NaiveTime, TimeZone,
+    Timelike, Utc,
+};
 
 use crate::decode::Decode;
 use crate::encode::{Encode, IsNull};
@@ -9,7 +12,7 @@ use crate::mysql::type_info::MySqlTypeInfo;
 use crate::mysql::{MySql, MySqlValueFormat, MySqlValueRef};
 use crate::types::Type;
 
-impl Type<MySql> for DateTime<Utc> {
+impl<T: TimeZone> Type<MySql> for DateTime<T> {
     fn type_info() -> MySqlTypeInfo {
         MySqlTypeInfo::binary(ColumnType::Timestamp)
     }
@@ -35,16 +38,6 @@ impl<'r> Decode<'r, MySql> for DateTime<Utc> {
     }
 }
 
-impl Type<MySql> for DateTime<Local> {
-    fn type_info() -> MySqlTypeInfo {
-        MySqlTypeInfo::binary(ColumnType::Timestamp)
-    }
-
-    fn compatible(ty: &MySqlTypeInfo) -> bool {
-        matches!(ty.r#type, ColumnType::Datetime | ColumnType::Timestamp)
-    }
-}
-
 /// Note: assumes the connection's `time_zone` is set to `+00:00` (UTC).
 impl Encode<'_, MySql> for DateTime<Local> {
     fn encode_by_ref(&self, buf: &mut Vec<u8>) -> IsNull {
@@ -56,6 +49,14 @@ impl Encode<'_, MySql> for DateTime<Local> {
 impl<'r> Decode<'r, MySql> for DateTime<Local> {
     fn decode(value: MySqlValueRef<'r>) -> Result<Self, BoxDynError> {
         Ok(<DateTime<Utc> as Decode<'r, MySql>>::decode(value)?.with_timezone(&Local))
+    }
+}
+
+/// Note: assumes the connection's `time_zone` is set to `+00:00` (UTC).
+impl<'r> Decode<'r, MySql> for DateTime<FixedOffset> {
+    fn decode(value: MySqlValueRef<'r>) -> Result<Self, BoxDynError> {
+        Ok(<DateTime<Utc> as Decode<'r, MySql>>::decode(value)?
+            .with_timezone(&FixedOffset::east_opt(0).unwrap()))
     }
 }